+++ Changelog.md
@@ -4,4 +4,5 @@
 -----
 
   * Add TreatAsLeaf attribute to mark properties to be treated as leafs even if they are not
+  * Use Pluralizer for nounes in ```PluralDataOwnerNameProvider```
   * Fixed the issue of undesired instantiations fot null properties in object.clone